refactor: move markers to user
