--- src/main/java/ru/practicum/shareit/marker/OnCreate.java
@@ -1,4 +1,0 @@
-package ru.practicum.shareit.marker;
-
-public interface OnCreate {
-}
