+++ src/main/java/ru/practicum/shareit/user/UserController.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-import ru.practicum.shareit.marker.OnCreate;
-import ru.practicum.shareit.marker.OnUpdate;
+import ru.practicum.shareit.user.marker.OnCreate;
+import ru.practicum.shareit.user.marker.OnUpdate;
 import ru.practicum.shareit.user.service.UserService;
 
 import java.util.List;
