+++ src/main/java/ru/practicum/shareit/user/UserDto.java
@@ -2,8 +2,8 @@ package ru.practicum.shareit.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import ru.practicum.shareit.marker.OnCreate;
-import ru.practicum.shareit.marker.OnUpdate;
+import ru.practicum.shareit.user.marker.OnCreate;
+import ru.practicum.shareit.user.marker.OnUpdate;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
