+++ src/main/java/ru/practicum/shareit/user/marker/OnUpdate.java
@@ -1,0 +1,4 @@
+package ru.practicum.shareit.user.marker;
+
+public interface OnUpdate {
+}
